Makefile, Kconfig: support for CROSS_COMPILE from kconfig.
The CROSS_COMPILE variable can by honnered by:
 * the command-line ($ make CROSS_COMPILE=arm-linux-gnueabi-) or by
 * the kconfig menu (CONFIG_CROSS_COMPILE=arm-none-linux-gnueabi- from .config)

The command-line takes precedence to .config.

Note that because the target architecture is always ARM, if no CROSS_COMPILE
variable is set, the Makefile will automatically set the arm-linux-gnueabi-
prefix.

To bypass this limitation, simply override the CROSS_COMPILE variable using the
command-line. For example, if you do not need to cross-compile because you are
compiling on an ARM architecture, you need to enter:

$ make CROSS_COMPILE=

But the compilation will be halted by the at91bootstrap. It will complains
about an empty CROSS_COMPILE variable.

Signed-off-by: Gaël PORTAY <gael.portay@gmail.com>